Create python created object collection
This collectionkeeps all references to ccObject created from python scripts
At the end of the execution of a script, the all objs in the collection are deleted.

When a user uses 'addToDB' to an obj that was created from Python into the cloud compare db tree
the obj is removed from the collection so that there is no double free.

The goal is to avoid memoryleak