Fix editor crashing after failing to load a file and loading another
If user tried to load a file and it failed (ie file did not exist)
and then tried to load another file, the editor would make CC crash regardless of whether the file exists.

It was because even when a file failed to load, its
CodeEditor would be added to the subWindowList and the next time a file would be loaded, the findChildCodeEditor function would make an incorect qobject_cast and then deref a nullptr.