Pick Up virtual environment CC is ran from
This adds the ability for the plugin to check if CloudCompare is being run from within a python virtual enviroment (python's venv or conda env) and configure its embeded python accordingly.

This is to allow users to create their own python environment where they can install  (pip/conda) extra dependencies and use them in their scripts.

Since the embeded has a version, this checks that the
venv version is exactly the same as the one the plugin was compiled with. 

If an error occurs, or the python version is not the same, warnings are printed to the ccConsole and the bundled python environement is used instead.

Also, if an error is caugh and that the config of the different paths could not be done correctly, the embed python is not initialized and all plugin actions are disabled.