Switch to using embedded modules
Use embeded modules for pycc & cccorelib.

This simplifies installing the Plugin as the modules are Bundled within the plugin, so we don't need to care about copying them at install and we don't need to
setup the python path at runtime to be able to load them. This should greatly simplify installation on unixes

Keep the ability to build them as independent modules
as it might be need later to be able to do 'pip installable modules'

(The cmake written to the changes might not be the best possible)