Make detachObject work with a Hack
Use a hack to make the detachObject mechanism of ccHObject work.

The idea is that function that return a child of a ccHObject
return a pybind11 object that is constructed to own the pointer, that means
it would delete the object when the ref count goes to 0. However we 'deactivate'
this mechanism by making it think it does not own the data.

And when the child object is detached from its parent we 'reactivate' the owning mechanism.

This is hack that I do not see how to handle in another way. And it may break one day