Allow user to select custom python env in settings
This commit adds in the settings a way for the users to select
a path to python env to be used by the plugin.

It also uses the PEP 587 feature to initialize the interpreter, and more importantly
handle errors when it could initialize itself properly as oposed to just exiting the program like before
allowing us to have a nicer user experience when something goes wrong with a user's local env.