PackageManager: Add checks on write rights
To be able to install a package, the package manager calls pip,
pip needs to have write permissions otherwise it will fail.

CloudCompare on windows being generally installed in C:\Programs
that requires admin rights, so a non admin cannot pip install things
in there.

There have been many user confused by this, so its easier to block
installation and have a message if we detect the folder is not writable

Also, the package manager should now automatically use the `--user` option
for pip install when it's using the system's Python