Fix crash after failing to init interpreter
If the interpreter failed to init, we did an early return
in the PythonPlugin constructor, leading to the QMenu to be null
resulting in the whole app crashing.

We fix that by still creating the menu
The getActions() function already properly enabled or not
the actions button depending on if the interpreter is initialized.
Meaning that now if the interpreter fails to initialize, then
CC will still open, but all python actions (expect for settings)
will be disabled.