Prevent invalid plugin dir from creating problems
In `PythonPluginManager::loadPluginsFrom`, a import of
pycc_runtime was done to get the subclasses of the plugins
base class.

However that import can only work if pycc had already been
imported before as pycc contains definitions necessary for pycc_runtime
to function. And plugins needs to import pycc to be able to subclass

So if no plugins were imported (either because of invalid path or empty
folders) the pycc_runtime import would fail, and subsequent import of
pycc_runtime would also fail making the runtime useless as no
interaction with the main app was possible

The fix is to import pycc intead as pycc will import pycc_runtime